refact : InvestStatusType 수정
PENDING, COMPLETED, CANCELLED에서 WATING, EXECUTING, COMPLETED, CANCELED으로 상태 분류 단계를 변경하고, description, detail 을 추가함

+++ src/main/java/com/billit/investment/repository/InvestmentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface InvestmentRepository extends JpaRepository<Investment, Long> {
+public interface InvestmentRepository extends JpaRepository<Investment, Integer> {
 
     List<Investment> findByUserInvestorId(Long userInvestorId);
 }
